requirements.txt
aiologger==0.7.0
disnake==2.9.1
python-dotenv
aiomysql
disnake-pagination